+++ src/nti/zope_catalog/tests/__init__.py
@@ -1,40 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from __future__ import print_function, unicode_literals, absolute_import, division
+from __future__ import print_function, absolute_import, division
 __docformat__ = "restructuredtext en"
 
-# disable: accessing protected members, too many methods
-# pylint: disable=W0212,R0904
+import zope.component.testlayer
+import nti.zope_catalog
 
-import unittest
-
-from nti.testing.layers import GCLayerMixin
-from nti.testing.layers import ZopeComponentLayer
-from nti.testing.layers import ConfiguringLayerMixin
-
-import zope.testing.cleanup
-
-
-class SharedConfiguringTestLayer(GCLayerMixin,
-                                 ZopeComponentLayer,
-                                 ConfiguringLayerMixin):
-
-    set_up_packages = ('nti.zope_catalog',)
-
-    @classmethod
-    def setUp(cls):
-        cls.setUpPackages()
-
-    @classmethod
-    def tearDown(cls):
-        cls.tearDownPackages()
-        zope.testing.cleanup.cleanUp()
-
-    @classmethod
-    def testSetUp(cls, test=None):
-        pass
-
-    @classmethod
-    def testTearDown(cls):
-        pass
+NTIZopeCatalogLayer = zope.component.testlayer.ZCMLFileLayer(nti.zope_catalog, 'configure.zcml')