+++ CHANGES.rst
@@ -7,3 +7,6 @@
 
 - First PyPI release.
 - Add support for Python 3.
+- ``TimestampNormalizer`` also normalizes incoming datetime objects.
+- Fix extent-based queries for NormalizedKeywordIndex.
+- 100% test coverage.